+++ client/.env
@@ -1,2 +1,3 @@
 VITE_APP_AUTH0_DOMAIN = "dev-xzd1nw16hc11vsj7.us.auth0.com"
 VITE_APP_AUTH0_CLIENT_ID = "joqKy6suHcPTxhs2CfgWWlTwKsMMSGyQ"
+VITE_APP_API_SERVER_URL = "http://localhost:8080"
