+++ client/.env.example
@@ -1,0 +1,11 @@
+# AUTH0 Environment Variables
+VITE_APP_AUTH0_DOMAIN = 
+VITE_APP_AUTH0_CLIENT_ID =
+VITE_APP_API_SERVER_URL = "http://localhost:8080/"
+#REACT_APP_AUTH0_DOMAIN=
+#REACT_APP_AUTH0_CLIENT_ID=
+#REACT_APP_AUTH0_CALLBACK_URL=http://localhost:4040/callback
+#REACT_APP_AUTH0_AUDIENCE=
+# Dirigirse a Auht0 -> https://manage.auth0.com/dashboard/us/dev-xzd1nw16hc11vsj7/applications 
+# para establacer las variables de entorno correctas
+# Crear un archivo .env y colocar las variables de entorno
